pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "phoenix-smartcaseai"
version = "1.0.0"
description = "AI-powered test case generation from user stories using LLMs"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "Phoenix SmartCase AI", email = "support@phoenixsmartcase.ai"},
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
keywords = [
    "test-case-generation",
    "bdd",
    "gherkin", 
    "ai",
    "llm",
    "openai",
    "testing",
    "qa",
    "user-stories",
    "automation"
]
dependencies = [
    "pydantic>=2.11.0,<3.0.0",
    "langchain-core>=0.3.76,<0.4.0",
    "langchain-openai>=0.3.30,<0.4.0",
    "langchain-google-genai>=2.0.0,<3.0.0",
    "langchain-anthropic>=0.3.0,<0.4.0",
    "python-docx>=0.8.11,<1.0.0",
    "pandas>=2.0.0,<3.0.0",
    "openpyxl>=3.1.0,<4.0.0",
    "PyPDF2>=3.0.0,<4.0.0",
    "pdfplumber>=0.9.0,<1.0.0",
    "Pillow>=10.0.0,<11.0.0",
    "pytesseract>=0.3.10,<1.0.0",
    "fastapi>=0.104.0,<1.0.0",
    "uvicorn[standard]>=0.24.0,<1.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "jinja2>=3.1.0,<4.0.0",
    "python-multipart>=0.0.6,<1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/shaktitrigent/Phoenix-SmartCaseAI"
Documentation = "https://github.com/shaktitrigent/Phoenix-SmartCaseAI#readme"
Repository = "https://github.com/shaktitrigent/Phoenix-SmartCaseAI.git"
Issues = "https://github.com/shaktitrigent/Phoenix-SmartCaseAI/issues"

[project.scripts]
phoenix-smartcase = "phoenix_smartcaseai.cli:main"
phoenix-smartcaseai = "phoenix_smartcaseai.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["phoenix_smartcaseai*"]
exclude = ["tests*"]

[tool.setuptools.package-data]
phoenix_smartcaseai = ["*.md", "*.txt", "templates/*"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = [
    "tests",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]
